Fix overflow caused by large timestamps
BUG=aomedia:2053

+++ common/ivfenc.c
@@ -22,15 +22,15 @@ void ivf_write_file_header(FILE *outfile, const struct aom_codec_enc_cfg *cfg,
   header[1] = 'K';
   header[2] = 'I';
   header[3] = 'F';
-  mem_put_le16(header + 4, 0);                     // version
-  mem_put_le16(header + 6, 32);                    // header size
-  mem_put_le32(header + 8, fourcc);                // fourcc
-  mem_put_le16(header + 12, cfg->g_w);             // width
-  mem_put_le16(header + 14, cfg->g_h);             // height
-  mem_put_le32(header + 16, cfg->g_timebase.den);  // rate
-  mem_put_le32(header + 20, cfg->g_timebase.num);  // scale
-  mem_put_le32(header + 24, frame_cnt);            // length
-  mem_put_le32(header + 28, 0);                    // unused
+  mem_put_le16(header + 4, 0);                          // version
+  mem_put_le16(header + 6, 32);                         // header size
+  mem_put_le32(header + 8, fourcc);                     // fourcc
+  mem_put_le16(header + 12, cfg->g_w);                  // width
+  mem_put_le16(header + 14, cfg->g_h);                  // height
+  mem_put_le32(header + 16, cfg->g_timebase.den);       // rate
+  mem_put_le32(header + 20, (int)cfg->g_timebase.num);  // scale
+  mem_put_le32(header + 24, frame_cnt);                 // length
+  mem_put_le32(header + 28, 0);                         // unused
 
   fwrite(header, 1, 32, outfile);
 }
